hack: remove is_x86_feature_detected!("bmi1")
It caused this error when compiling with the Teaclave Rust SGX SDK v1.1.3
error: unknown cpu feature: bmi1
This is because of a shortcoming in the SDK, which doesn't support
recently added x86_features in new versions of the Rust std library

Warning : do not use in production, this is a hack to circumvent
compilation problem. This check should be done.